fix: Fixed waiting on update module to count updates (#82)
* fix: Fixed bugs in hyprlock not launching

* fix: Fixed waiting on update module to count updates

* fix: Fixed waiting on update module to count updates

* fix: updated using suggested change

* fix: updated using env file and removed jquery

* fix: used GLOBAL variable convention

---------

+++ Configs/.local/lib/hyde/systemupdate.sh
@@ -7,27 +7,46 @@ fi
 
 # source variables
 scrDir=$(dirname "$(realpath "$0")")
+# shellcheck disable=SC1091
 source "$scrDir/globalcontrol.sh"
 get_aurhlpr
 export -f pkg_installed
 fpk_exup="pkg_installed flatpak && flatpak update"
+temp_file="$HYDE_RUNTIME_DIR/update_info"
+# shellcheck source=/dev/null
+[ -f "$temp_file" ] && source "$temp_file"
 
 # Trigger upgrade
 if [ "$1" == "up" ] ; then
-    trap 'pkill -RTMIN+20 waybar' EXIT
-    command="
-    fastfetch
-    $0 upgrade
-    ${aurhlpr} -Syu
-    $fpk_exup
-    read -n 1 -p 'Press any key to continue...'
-    "
-    kitty --title systemupdate sh -c "${command}"
+    if [ -f "$temp_file" ]; then
+        # refreshes the module so after you update it will reset to zero
+        trap 'pkill -RTMIN+20 waybar' EXIT
+        # Read info from env file
+        while IFS="=" read -r key value; do
+            case "$key" in
+                OFFICIAL_UPDATES) official=$value ;;
+                AUR_UPDATES) aur=$value ;;
+                FLATPAK_UPDATES) flatpak=$value ;;
+            esac
+        done < "$temp_file"
+
+        command="
+        fastfetch
+        printf '[Official] %-10s\n[AUR]      %-10s\n[Flatpak]  %-10s\n' '$official' '$aur' '$flatpak'
+        ${aurhlpr} -Syu
+        $fpk_exup
+        read -n 1 -p 'Press any key to continue...'
+        "
+        kitty --title systemupdate sh -c "${command}"
+    else
+        echo "No upgrade info found. Please run the script without parameters first."
+    fi
+    exit 0
 fi
 
 # Check for AUR updates
-aur=$(${aurhlpr} -Qua | wc -l)
-ofc=$( (while pgrep -x checkupdates > /dev/null ; do sleep 1; done) ; checkupdates | wc -l)
+aur=$(${aurhlpr} -Qua | wc -l) 
+ofc=$(pacman -Qu | wc -l)
 
 # Check for flatpak updates
 if pkg_installed flatpak ; then
@@ -40,9 +59,16 @@ fi
 
 # Calculate total available updates
 upd=$(( ofc + aur + fpk ))
+# Prepare the upgrade info
+upgrade_info=$(cat <<EOF
+OFFICIAL_UPDATES=$ofc
+AUR_UPDATES=$aur
+FLATPAK_UPDATES=$fpk
+EOF
+)
 
-[ "${1}" == upgrade ] && printf "[Official] %-10s\n[AUR]      %-10s\n[Flatpak]  %-10s\n" "$ofc" "$aur" "$fpk" && exit
-
+# Save the upgrade info
+echo "$upgrade_info" > "$temp_file"
 # Show tooltip
 if [ $upd -eq 0 ] ; then
     upd="" #Remove Icon completely
